--- a/Cargo.toml
+++ b/Cargo.toml
@@ -27,8 +27,10 @@ walkdir = "2.3.2"
 
 [dev-dependencies]
 criterion = "0.4.0"
+env_logger = "0.10.0"
 insta = { version = "1.22.0", features = ["yaml", "redactions"] }
 regex = "1.7.0"
+test-log = "0.2.11"
 
 [profile.dev.package.insta]
 opt-level = 3