--- a/src/solver/backends.rs
+++ b/src/solver/backends.rs
@@ -4,6 +4,7 @@
 use std::{
     collections::{HashMap, HashSet},
     iter::zip,
+    time::Instant,
 };
 
 use crate::{
@@ -15,10 +16,11 @@ use crate::{
         proc::{CvcConf, SolverCmd, Z3Conf},
         sexp::{self, atom_s},
     },
+    solver::models::ModelSymbol,
 };
 
 use super::{
-    models::{self, subst},
+    models::{self, subst, PartialInterp},
     {Backend, FOModel},
 };
 
@@ -97,27 +99,30 @@ impl Backend for &GenericBackend {
             .map(|(sort, elements)| (sort.clone(), elements.len()))
             .collect();
 
-        let mut interps = HashMap::new();
-        for (symbol, (binders, body)) in &model.symbols {
+        let mut part_interp = PartialInterp::for_model(&model);
+        for (symbol, model_sym) in &model.symbols {
             if indicators.contains(symbol) {
                 continue;
             }
-            let symbol_no_primes = symbol.trim_end_matches(|c| c == '\'');
-            // this symbol is not in the signature
-            if !sig.relations.iter().any(|r| r.name == symbol_no_primes) {
-                continue;
-            }
-            let rel = sig.relation_decl(symbol);
-            let mut shape = rel
-                .args
+            let start = Instant::now();
+            let ModelSymbol {
+                binders,
+                body,
+                ret_sort,
+            } = model_sym;
+            let arg_sorts = binders
+                .iter()
+                .map(|(_, sort)| sort.clone())
+                .collect::<Vec<_>>();
+            let mut shape = arg_sorts
                 .iter()
                 .map(|sort| sort_cardinality(&universe, sort))
                 .collect::<Vec<usize>>();
-            shape.push(sort_cardinality(&universe, &rel.sort));
+            shape.push(sort_cardinality(&universe, ret_sort));
             let interp = Interpretation::new(&shape, |args: &[Element]| -> Element {
                 // get the arguments as terms, based on model.universes
-                let args = zip(args, &rel.args)
-                    .map(|(&e_idx, sort)| match sort {
+                let args = zip(args, &arg_sorts)
+                    .map(|(&e_idx, typ)| match typ {
                         Sort::Bool => {
                             if e_idx == 0 {
                                 atom_s("false")
@@ -132,13 +137,13 @@ impl Backend for &GenericBackend {
                         }
                     })
                     .collect::<Vec<_>>();
-                let repl = zip(binders.iter().map(|s| s.as_str()), args).collect::<Vec<_>>();
+                let repl = zip(binders.iter().map(|s| s.0.as_str()), args).collect::<Vec<_>>();
                 let body = subst(&repl, body);
                 let e = model
-                    .smt_eval(&body)
-                    .unwrap_or_else(|err| panic!("could not interpret {}: {err}", &rel.name));
+                    .smt_eval(&part_interp, &body)
+                    .unwrap_or_else(|err| panic!("could not interpret {symbol}: {err}"));
                 let res = e.s().expect("unhandled evaluation result");
-                match &rel.sort {
+                match ret_sort {
                     Sort::Bool => {
                         if res == "false" {
                             return 0;
@@ -158,13 +163,29 @@ impl Backend for &GenericBackend {
                     }
                 }
             });
-            interps.insert(symbol.clone(), interp);
+            part_interp
+                .interps
+                .insert(symbol.clone(), (interp, ret_sort.clone()));
+            log::debug!(
+                "evaluated {symbol} in {:.1}s",
+                Instant::now().duration_since(start).as_secs_f64()
+            )
         }
 
-        FOModel {
-            universe,
-            interp: interps,
-        }
+        let interp = part_interp
+            .interps
+            .into_iter()
+            .filter(|(symbol, _)| {
+                let symbol_no_primes = symbol.trim_end_matches(|c| c == '\'');
+                // this symbol is not in the signature
+                if !sig.relations.iter().any(|r| r.name == symbol_no_primes) {
+                    return false;
+                }
+                return true;
+            })
+            .map(|(symbol, (interp, _))| (symbol, interp))
+            .collect();
+        FOModel { universe, interp }
     }
 }
 
@@ -176,8 +197,10 @@ mod tests {
 
     use super::{Backend, GenericBackend, SolverType};
 
+    use test_log::test;
+
     #[test]
-    #[ignore] // parsing this model takes too long
+    #[ignore]
     fn test_issue_5_parse_model_with_auxilliary_defs() {
         let _ = pretty_env_logger::try_init();
         let sig = parser::parse_signature(