--- a/src/solver/models.rs
+++ b/src/solver/models.rs
@@ -8,12 +8,33 @@ use thiserror::Error;
 
 use regex::Regex;
 
-use crate::smtlib::sexp::{atom_s, sexp_l, Atom, Sexp};
+use crate::{
+    fly::{
+        semantics::{Element, Interpretation},
+        syntax::Sort,
+    },
+    smtlib::sexp::{atom_s, sexp_l, Atom, Sexp},
+};
+
+#[derive(Debug, Clone)]
+pub struct PartialInterp {
+    universes: HashMap<String, Vec<String>>,
+    // reverse of universes
+    term_to_element: HashMap<String, Element>,
+    pub interps: HashMap<String, (Interpretation, Sort)>,
+}
+
+#[derive(Debug, Clone, Serialize, PartialEq, Eq)]
+pub struct ModelSymbol {
+    pub binders: Vec<(String, Sort)>,
+    pub body: Sexp,
+    pub ret_sort: Sort,
+}
 
 #[derive(Debug, Clone, Serialize, PartialEq, Eq)]
 pub struct Model {
     pub universes: HashMap<String, Vec<String>>,
-    pub symbols: HashMap<String, (Vec<String>, Sexp)>,
+    pub symbols: HashMap<String, ModelSymbol>,
 }
 
 #[derive(Debug, Error)]
@@ -43,7 +64,16 @@ pub fn subst(repl: &[(&str, Sexp)], e: &Sexp) -> Sexp {
     subst_hashmap(&repl_map, e)
 }
 
-fn parse_binders(binders: &Sexp) -> Vec<String> {
+fn parse_sort(sort: &Sexp) -> Sort {
+    let sort_name = sort.atom_s().unwrap();
+    if sort_name == "Bool" {
+        Sort::Bool
+    } else {
+        Sort::Id(sort_name.to_string())
+    }
+}
+
+fn parse_binders(binders: &Sexp) -> Vec<(String, Sort)> {
     let binder_sexps = binders.list().unwrap();
     binder_sexps
         .iter()
@@ -51,14 +81,16 @@ fn parse_binders(binders: &Sexp) -> Vec<String> {
             // b should be a list of (name type)
             let ss = b.list().unwrap();
             assert!(ss.len() == 2, "binder is ill-formed");
-            ss[0].atom_s().unwrap().to_string()
+            let name = ss[0].atom_s().unwrap().to_string();
+            let sort = parse_sort(&ss[1]);
+            (name, sort)
         })
         .collect()
 }
 
 pub(crate) fn parse_z3(model: &Sexp) -> Model {
     let mut universes: HashMap<String, Vec<String>> = HashMap::new();
-    let mut symbols: HashMap<String, (Vec<String>, Sexp)> = HashMap::new();
+    let mut symbols: HashMap<String, ModelSymbol> = HashMap::new();
     if let Some(ss) = model.list() {
         // remove a leading "model" for older versions
         let ss = if !ss.is_empty() && ss[0] == atom_s("model") {
@@ -88,9 +120,14 @@ pub(crate) fn parse_z3(model: &Sexp) -> Model {
                     );
                     let name = args[0].atom_s().unwrap().to_string();
                     let binders = parse_binders(&args[1]);
-                    // let sort = args[2].atom_s().unwrap();
+                    let ret_sort = parse_sort(&args[2]);
                     let body = args[3].clone();
-                    symbols.insert(name, (binders, body));
+                    let sym = ModelSymbol {
+                        binders,
+                        body,
+                        ret_sort,
+                    };
+                    symbols.insert(name, sym);
                 } else if head == "forall" {
                     // ignore, cardinality constraint
                 } else {
@@ -105,7 +142,7 @@ pub(crate) fn parse_z3(model: &Sexp) -> Model {
 pub(crate) fn parse_cvc(model: &Sexp, version5: bool) -> Model {
     let version4 = !version5;
     let mut universe_cardinalities: HashMap<String, usize> = HashMap::new();
-    let mut symbols: HashMap<String, (Vec<String>, Sexp)> = HashMap::new();
+    let mut symbols: HashMap<String, ModelSymbol> = HashMap::new();
     lazy_static! {
         static ref CARDINALITY_RE: Regex = Regex::new("cardinality of (.*) is ([0-9]+)$").unwrap();
     }
@@ -138,9 +175,14 @@ pub(crate) fn parse_cvc(model: &Sexp, version5: bool) -> Model {
                     );
                     let name = args[0].atom_s().unwrap().to_string();
                     let binders = parse_binders(&args[1]);
-                    // let sort = args[2].atom_s().unwrap();
+                    let ret_sort = parse_sort(&args[2]);
                     let body = args[3].clone();
-                    symbols.insert(name, (binders, body));
+                    let sym = ModelSymbol {
+                        binders,
+                        body,
+                        ret_sort,
+                    };
+                    symbols.insert(name, sym);
                 } else if version4 && head == "declare-sort" {
                     // cvc4 only
                     continue;
@@ -166,9 +208,61 @@ pub(crate) fn parse_cvc(model: &Sexp, version5: bool) -> Model {
     Model { universes, symbols }
 }
 
+impl PartialInterp {
+    pub fn for_model(model: &Model) -> Self {
+        let mut term_to_element = HashMap::new();
+        // the sort doesn't matter here (universe element names are already
+        // distinct between sorts)
+        for elements in model.universes.values() {
+            for (e_idx, term) in elements.iter().enumerate() {
+                term_to_element.insert(term.to_string(), e_idx);
+            }
+        }
+        Self {
+            term_to_element,
+            universes: model.universes.clone(),
+            interps: HashMap::new(),
+        }
+    }
+
+    fn has_eval(&self, f: &str) -> bool {
+        self.interps.contains_key(f)
+    }
+
+    fn eval(&self, f: &str, args: &[Atom]) -> String {
+        let (interp, ret_sort) = &self.interps[f];
+        let args = args
+            .iter()
+            .map(|atom| match atom {
+                Atom::I(_) => panic!("cannot evaluate on integers"),
+                Atom::S(name) => {
+                    if name == "true" {
+                        1
+                    } else if name == "false" {
+                        0
+                    } else {
+                        self.term_to_element[name]
+                    }
+                }
+            })
+            .collect::<Vec<usize>>();
+        let result_el = interp.get(&args);
+        match ret_sort {
+            Sort::Bool => {
+                if result_el == 1 {
+                    "true".to_string()
+                } else {
+                    "false".to_string()
+                }
+            }
+            Sort::Id(sort) => self.universes[sort][result_el].clone(),
+        }
+    }
+}
+
 impl Model {
-    fn eval_bool(&self, e: &Sexp) -> Result<bool, EvalError> {
-        let a = self.smt_eval(e)?;
+    fn eval_bool(&self, part_eval: &PartialInterp, e: &Sexp) -> Result<bool, EvalError> {
+        let a = self.smt_eval(part_eval, e)?;
         match a {
             Atom::S(s) if s == "true" => Ok(true),
             Atom::S(s) if s == "false" => Ok(false),
@@ -178,7 +272,10 @@ impl Model {
 
     /// Evaluate an SMT expression, reducing constants with known semantics. Fails
     /// if this does not result in an Atom.
-    pub fn smt_eval(&self, e: &Sexp) -> Result<Atom, EvalError> {
+    pub fn smt_eval(&self, part_eval: &PartialInterp, e: &Sexp) -> Result<Atom, EvalError> {
+        // println!("evaluating {e}");
+        let go_bool = |e: &Sexp| self.eval_bool(part_eval, e);
+        let go = |e: &Sexp| self.smt_eval(part_eval, e);
         match e {
             Sexp::Atom(a) => Ok(a.clone()),
             Sexp::Comment(_) => Err(EvalError("comment".to_string())),
@@ -198,36 +295,36 @@ impl Model {
                 if head == "and" {
                     let args = args
                         .iter()
-                        .map(|s| self.eval_bool(s))
+                        .map(|s| go_bool(s))
                         .collect::<Result<Vec<_>, _>>()?;
                     let v = args.iter().all(|x| *x);
                     Ok(bool_atom(v))
                 } else if head == "or" {
                     let args = args
                         .iter()
-                        .map(|s| self.eval_bool(s))
+                        .map(|s| go_bool(s))
                         .collect::<Result<Vec<_>, _>>()?;
                     let v = args.iter().any(|x| *x);
                     Ok(bool_atom(v))
                 } else if head == "not" || head == "!" {
                     assert_eq!(args.len(), 1);
-                    let x = self.eval_bool(args[0])?;
+                    let x = go_bool(args[0])?;
                     Ok(bool_atom(!x))
                 } else if head == "as" {
                     // type cast (basically ignored)
-                    self.smt_eval(args[0])
+                    go(args[0])
                 } else if head == "=" {
                     assert_eq!(args.len(), 2);
-                    let lhs = self.smt_eval(args[0])?;
-                    let rhs = self.smt_eval(args[1])?;
+                    let lhs = go(args[0])?;
+                    let rhs = go(args[1])?;
                     Ok(bool_atom(lhs == rhs))
                 } else if head == "ite" {
                     assert_eq!(args.len(), 3);
-                    let cond = self.eval_bool(args[0])?;
+                    let cond = go_bool(args[0])?;
                     if cond {
-                        self.smt_eval(args[1])
+                        go(args[1])
                     } else {
-                        self.smt_eval(args[2])
+                        go(args[2])
                     }
                 } else if head == "let" {
                     // (let ((x1 e1) (x2 e2)) e)
@@ -244,18 +341,30 @@ impl Model {
                         })
                         .collect();
                     let e = subst(&repl, args[1]);
-                    self.smt_eval(&e)
+                    go(&e)
                 } else if self.symbols.contains_key(head) {
-                    let (binders, body) = &self.symbols[head];
+                    if part_eval.has_eval(head) {
+                        let args = args.iter().flat_map(|e| go(e)).collect::<Vec<_>>();
+                        let res = part_eval.eval(head, &args);
+                        return Ok(Atom::S(res));
+                    }
+                    let ModelSymbol { binders, body, .. } = &self.symbols[head];
                     assert_eq!(
                         binders.len(),
                         args.len(),
                         "wrong number of arguments to {head}"
                     );
+                    let args = args
+                        .iter()
+                        .flat_map(|e| -> Result<Sexp, EvalError> {
+                            let a = go(e)?;
+                            Ok(Sexp::Atom(a))
+                        })
+                        .collect::<Vec<_>>();
                     let repl: Vec<(&str, Sexp)> = iter::zip(binders, args.iter().cloned())
-                        .map(|(name, e)| (name.as_str(), e.clone()))
+                        .map(|(binder, e)| (binder.0.as_str(), e))
                         .collect();
-                    self.smt_eval(&subst(&repl, body))
+                    go(&subst(&repl, body))
                 } else {
                     Err(EvalError(format!("unexpected function {head}")))
                 }