Model evaluation could be very slow for Z3 models because it sometimes uses large auxilliary functions, which are evaluated over and over again while enumerating all the inputs for other functions. This is exacerbated by the fact that its models aren't minimized, so the universes are large - see #37 for some notes on that.

In order to speed up evaluation, this PR first constructs an `Interpretation` for every model symbol (not just the ones in the signature), and then it evaluates calls to previously evaluated functions using their `Interpretation` rather than by substituting their definition.

I added a new test which is just one auxilliary definition and one additional symbol, extracted from the large model. This test used to take 13s and now takes under 0.5s (both in release mode). Even in a debug build (which is the default for tests) this test takes 2s, which is acceptable for a regularly-run unit test.

The old test does work but takes over 10s in debug mode, so I've left it ignored for now. You can see the performance in release mode by running

```sh
cargo test --release -- --ignored parse_model_with_auxilliary_defs
```

This PR isn't fully ready because the model symbols aren't evaluated in the right order yet. In order to get any performance benefit, a symbol has to be interpreted before it's used in other definitions, which requires evaluating in topological order. In order to make this code work at all I added a hack that sorts the auxilliary definitions first (those that aren't in the signature and only in the model), which makes these examples fast and deterministic compared to HashMap iteration order. Ultimately we should evaluate on demand, though, which may even be slightly faster since Z3 does use even other symbols in the signature in its models.

Fixes #26.